master: move load_config and run to app module
